refactor: Remove superadmin controllers and add global exception handling
- Deleted multiple superadmin controllers for various entities
- Added GlobalExceptionHandler for centralized error handling
- Updated some entity classes with additional fields
- Introduced new authentication and captcha-related service classes

+++ src/main/java/com/sakurapuare/boatmanagement/pojo/entity/Accounts.java
@@ -5,7 +5,6 @@ import com.mybatisflex.annotation.KeyType;
 import com.mybatisflex.annotation.Table;
 import com.sakurapuare.boatmanagement.pojo.entity.BaseEntity;
 import java.io.Serializable;
-import java.math.BigInteger;
 
 import java.io.Serial;
 
@@ -21,7 +20,7 @@ import lombok.EqualsAndHashCode;
  * 基础账号表 实体类。
  *
  * @author sakurapuare
- * @since 2025-02-08
+ * @since 2025-02-10
  */
 @Data
 @Builder
@@ -40,7 +39,7 @@ public class Accounts extends BaseEntity implements Serializable {
      */
     @Id(keyType = KeyType.Auto)
     @ApiModelProperty("主键ID")
-    private BigInteger id;
+    private Long id;
 
     /**
      * 用户名
