+++ src/main/java/com/sakurapuare/boatmanagement/pojo/entity/CaptchaLimit.java
@@ -5,7 +5,6 @@ import com.mybatisflex.annotation.KeyType;
 import com.mybatisflex.annotation.Table;
 import com.sakurapuare.boatmanagement.pojo.entity.BaseEntity;
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.sql.Timestamp;
 
 import java.io.Serial;
@@ -22,7 +21,7 @@ import lombok.EqualsAndHashCode;
  * 验证码防刷记录 实体类。
  *
  * @author sakurapuare
- * @since 2025-02-08
+ * @since 2025-02-10
  */
 @Data
 @Builder
@@ -38,7 +37,7 @@ public class CaptchaLimit extends BaseEntity implements Serializable {
 
     @Id(keyType = KeyType.Auto)
     @ApiModelProperty("")
-    private BigInteger id;
+    private Long id;
 
     @ApiModelProperty("")
     private String target;
