+++ src/main/java/com/sakurapuare/boatmanagement/pojo/entity/SocialAuth.java
@@ -5,7 +5,6 @@ import com.mybatisflex.annotation.KeyType;
 import com.mybatisflex.annotation.Table;
 import com.sakurapuare.boatmanagement.pojo.entity.BaseEntity;
 import java.io.Serializable;
-import java.math.BigInteger;
 
 import java.io.Serial;
 
@@ -21,7 +20,7 @@ import lombok.EqualsAndHashCode;
  * 第三方登录表 实体类。
  *
  * @author sakurapuare
- * @since 2025-02-08
+ * @since 2025-02-10
  */
 @Data
 @Builder
@@ -37,10 +36,10 @@ public class SocialAuth extends BaseEntity implements Serializable {
 
     @Id(keyType = KeyType.Auto)
     @ApiModelProperty("")
-    private BigInteger id;
+    private Long id;
 
     @ApiModelProperty("")
-    private BigInteger accountId;
+    private Long accountId;
 
     /**
      * 第三方平台
